refactor(toolkit): internal deepening batch — headless context factory, wrapper DOM-snapshot module, utilities split
Collapse the five duplicated headless DI resolutions (config, form
context, messages registry, label resolver) into one private
buildHeadlessContext() factory; move the wrapper's afterEveryRender
DOM-snapshot protocol into a private module with a single
captureFormFieldWrapperDomSnapshot() entry point and consolidate the
selection-cluster ARIA wiring into one resolveClusterAriaAttrs()
computed; split headless utilities.ts by domain into
field-state-utilities.ts and error-summary-utilities.ts behind the
unchanged public barrel.

Internal only: no barrel file changes, published surface verified by
packaging.spec, DOM/ARIA output unchanged.

Closes #354
