+++ packages/toolkit/headless/src/lib/create-error-message-signal.ts
@@ -1,12 +1,9 @@
-import { computed, inject, type Injector, type Signal } from '@angular/core';
+import { computed, type Injector, type Signal } from '@angular/core';
 import type { ValidationError } from '@angular/forms/signals';
-import { NGX_SIGNAL_FORMS_CONFIG } from '@ngx-signal-forms/toolkit';
 import {
   assertInjector,
-  DEFAULT_NGX_SIGNAL_FORMS_CONFIG,
   createErrorVisibility,
   generateErrorId,
-  NGX_ERROR_MESSAGES,
   readDirectErrors,
   splitByKind,
   stripAngularFormPrefix,
@@ -16,6 +13,7 @@ import {
   type SubmittedStatus,
 } from '@ngx-signal-forms/toolkit/core';
 
+import { buildHeadlessContext } from './build-headless-context';
 import { resolveErrorMessage } from './utilities';
 
 /**
@@ -224,28 +222,25 @@ export function createErrorMessageSignal(
 ): Signal<readonly ResolvedFieldError[]> {
   return assertInjector(createErrorMessageSignal, options?.injector, () => {
     const registrySignal = options?.errorMessages;
-    // Only auto-inject when no explicit registry is supplied. Calling
-    // `inject()` when a registry is already provided would force callers to
-    // wrap every test in `runInInjectionContext`, defeating the explicit
-    // override.
-    const injectedRegistry =
-      registrySignal === undefined
-        ? inject(NGX_ERROR_MESSAGES, { optional: true })
-        : null;
-
-    // Falls back to the global `defaultErrorStrategy` config (same cascade
-    // `NgxHeadlessFieldset` applies) when neither an explicit `strategy` nor
-    // a form context is present, keeping standalone usage consistent
-    // regardless of which headless surface a consumer reaches for.
-    const config =
-      inject(NGX_SIGNAL_FORMS_CONFIG, { optional: true }) ??
-      DEFAULT_NGX_SIGNAL_FORMS_CONFIG;
+    // Resolves the shared ambient DI contract (form context, config,
+    // registry, label resolver) in one call. `buildHeadlessContext()` always
+    // resolves `NGX_ERROR_MESSAGES` — even when an explicit `errorMessages`
+    // registry is supplied and the resolved value goes unused below — since
+    // an optional token resolution is inexpensive and every other headless
+    // call site already resolves the full contract unconditionally the same
+    // way.
+    const { config, errorMessagesRegistry: injectedRegistry } =
+      buildHeadlessContext();
 
     const visibilityOptions: CreateErrorVisibilityOptions = {
       ...(options?.strategy !== undefined && { strategy: options.strategy }),
       ...(options?.submittedStatus !== undefined && {
         submittedStatus: options.submittedStatus,
       }),
+      // Falls back to the global `defaultErrorStrategy` config (same cascade
+      // `NgxHeadlessFieldset` applies) when neither an explicit `strategy` nor
+      // a form context is present, keeping standalone usage consistent
+      // regardless of which headless surface a consumer reaches for.
       ...(config?.defaultErrorStrategy !== undefined && {
         configDefault: config.defaultErrorStrategy,
       }),
